+++ tests/test_projection_pushdown.py
@@ -363,7 +363,6 @@ class TestProjectionPushdown:
 
         vcf_read_options = VcfReadOptions(
             info_fields=None,
-            thread_num=1,
             object_storage_options=object_storage_options,
         )
         read_options = ReadOptions(vcf_read_options=vcf_read_options)